+++ CHANGELOG.md
@@ -10,7 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed review stars laoding
 
+### Added
+
+- Updated the App name and using the graphql from common folder
+- Updated cy-runner.yml and removed unneccesary selectors
+
 ## [3.12.1] - 2022-12-05
+
 
 ### Added
 
@@ -21,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Italian and English translations.
 
 ## [3.12.0] - 2022-11-04
+
 
 ### Added